Support tracking of different client metrics independently

+++ ocelli-core/src/main/java/netflix/ocelli/metrics/math/Average.java
@@ -3,6 +3,6 @@ package netflix.ocelli.metrics.math;
 
 public interface Average {
     double get();
-    void addSample(int sample);
+    void addSample(long sample);
     void reset();
 }
